Support unicode outside of BMP.
`.length` and `.charAt` iterate over UTF-16 code points, and not actual
characters.  So for characters outside of the UTF-16 range (Basic
Multilingual Plane), this would split them in half and result in broken
characters (and broken in a way that we don't want ;p).

Use ES6 `for...of` loop to correctly loop over the characters and leave
the surrogate pairs together.